requirements.txt
python-dotenv
google-generativeai
numpy
faiss-cpu
sentence-transformers